fix regex used in evil-tex-toggle-delim
Previously, it would match anything that had a slash followed optionally by a
balancing keyword followed optionally by l, which would mess up how it matched
things like \{ or \|

Currently, the regex it uses to decide if the delimiters are already
balanced (based just on the initial one) is:

"\\\\\\(?:left\\|big\\|bigg\\|Big\\|Bigg\\)?l?"

(note: all regexes here are taken from elisp so have extra backslashes for
string escapes)

The trouble is that this matches *anything* that has a slash, since everything
that follows is optional. For example, if you have math like $ \{ x \} $ and you
run evil-tex-toggle-delim, then this *will* match (just the literal backslash),
causing it to think that it is already size balanced, and then "toggling" will
just strip the backslash rather than prepending \left and \right.

This could be fixed by changing it to

"\\\\\\(?:\\(left\\|big\\|bigg\\|Big\\|Bigg\\)\\|l\\)"

This extra level of grouping means that it only matches if there is a literal
backslash followed by *either* one of the balancing keywords OR l. This way it
has to match something in the second part, and this will not match \{

However, I think that trailing (optional) l is perhaps a mistake. I assume it is
there to catch things like \langle, but this will then toggle \langle to
angle (which without the slash is not a macro). Moreover, \langle and
\rangle (or similarly \lvert and \rvert, etc) are paired delimiters but they do
not automatically resize (at least not for me), I instead need to do
\left\langle if I want that behavior. For that reason, I think the trailing l
should be cut, and the final regex is just

"\\\\\\(?:left\\|big\\|bigg\\|Big\\|Bigg\\)"